add unit test

+++ SitNBeer/src/main/java/com/example/demo/controllers/BarController.java
@@ -108,7 +108,7 @@ class BarController {
 	}
 
 	@GetMapping("/bar/{bar_id}")
-	public String showBar(Model model, @PathVariable Integer bar_id){
+	public String showBar(Model model, @PathVariable long bar_id){
 		Optional<Bar> optionalBar = barRepository.findById(bar_id);
 		Bar bar = null;
 		if(optionalBar.isPresent()){
